feat(settings): Change default screenshot colours
