--- src/@types/screenshot.ts
@@ -1,5 +1,0 @@
-export enum ScreenshotBgStyle {
-  Transparent = "transparent",
-  Color = "color",
-  Theme = "theme",
-}
