+++ src/main.ts
@@ -1,48 +1,13 @@
-import { Plugin, Vault, Workspace, WorkspaceLeaf } from "obsidian";
+import { Plugin as ObsidianPlugin, Vault, Workspace, WorkspaceLeaf } from "obsidian";
 
-import MindmapView from "./mindmap-view";
+import View from "./mindmap-view";
 import { MM_VIEW_TYPE } from "./constants";
-import { MindMapSettingsTab } from "./settings-tab";
+import { SettingsTab } from "./settings-tab";
 import { inlineRenderer } from "./inline-renderer";
 
-import { ScreenshotBgStyle } from "./@types/screenshot";
+import { DEFAULT_SETTINGS, MindMapSettings, ScreenshotBgStyle } from "./@types/settings";
 
-const DEFAULT_SETTINGS: MindMapSettings = {
-  splitDirection: "horizontal",
-  nodeMinHeight: 16,
-  lineHeight: "1em",
-  spacingVertical: 5,
-  spacingHorizontal: 80,
-  paddingX: 8,
-  color1: "#fed766",
-  color1Thickness: "8",
-
-  color2: "#2ab7ca",
-  color2Thickness: "6",
-
-  color3: "#fe4a49",
-  color3Thickness: "4",
-
-  defaultColor: "#000",
-  defaultColorThickness: "2",
-
-  initialExpandLevel: -1,
-
-  coloring: "depth",
-
-  colorFreezeLevel: 0,
-  animationDuration: 500,
-  maxWidth: 0,
-  // below a beautiful dark blue color
-  screenshotBgColor: "#039614",
-  screenshotBgStyle: ScreenshotBgStyle.Transparent,
-  screenshotTransparentBg: true,
-  screenshotTextColor: "#ffffff",
-  screenshotTextColorEnabled: false,
-  highlight: true,
-};
-
-export default class MindMap extends Plugin {
+export default class Plugin extends ObsidianPlugin {
   vault: Vault;
   workspace: Workspace;
   settings: MindMapSettings;
@@ -51,14 +16,11 @@ export default class MindMap extends Plugin {
     console.log("Loading Mind Map plugin");
     this.vault = this.app.vault;
     this.workspace = this.app.workspace;
-    this.settings = Object.assign(
-      { ...DEFAULT_SETTINGS },
-      await this.loadData()
-    );
+    this.settings = { ...DEFAULT_SETTINGS, ...await this.loadData() };
 
     this.registerView(
       MM_VIEW_TYPE,
-      (leaf: WorkspaceLeaf) => new MindmapView(this.settings, leaf)
+      (leaf: WorkspaceLeaf) => new View(this.settings, leaf)
     );
 
     this.addCommand({
@@ -70,7 +32,7 @@ export default class MindMap extends Plugin {
       hotkeys: [],
     });
 
-    this.addSettingTab(new MindMapSettingsTab(this.app, this));
+    this.addSettingTab(new SettingsTab(this.app, this));
 
     this.registerMarkdownCodeBlockProcessor(
       "markmap",
@@ -90,7 +52,7 @@ export default class MindMap extends Plugin {
       "split",
       this.settings.splitDirection
     );
-    const mmPreview = new MindmapView(this.settings, preview);
+    const mmPreview = new View(this.settings, preview);
 
     preview.open(mmPreview);
   }
